added a fake alpha test for selecting obstacles
i made it so the larger the collider box is on an obstacle, the further
away the collider box is away from the camera is. so a smaller object
inside a larger object can be grabbed if you click on it. this will
"feel" like it depends on the alpha of an object when it actually does
not